+++ rust/tw_bitcoin/src/modules/mod.rs
@@ -1,7 +1,3 @@
 pub mod legacy;
-mod signer;
-mod transactions;
-
-// Re-exports
-pub use signer::Signer;
-pub use transactions::*;
+pub mod signer;
+pub mod transactions;
