+++ rust/tw_bitcoin/src/tests/address.rs
@@ -1,13 +1,12 @@
-use crate::{Result, Recipient};
+use crate::{Recipient, Result};
 use bitcoin::key::KeyPair;
-use bitcoin::{Network, PublicKey, PrivateKey};
+use bitcoin::{Network, PrivateKey, PublicKey};
 
 pub fn keypair_from_wif(string: &str) -> Result<KeyPair> {
     let pk = PrivateKey::from_wif(string).unwrap();
     let keypair = KeyPair::from_secret_key(&secp256k1::Secp256k1::new(), &pk.inner);
     Ok(keypair)
 }
-
 
 // This private key was used in a Bitcoin regtest environment.
 pub const ALICE_WIF: &str = "cQUNzeMnF9xPPLqZhH7hMVYGwSuu3b78zznuc5UrxgXnYQBq6Bx1";
